# Speedup localgraph