I have done my tests & small optimisation on LocalGraph #1338 .

So, to sum up :
 - Replace seen unordered_set by a boolean check.
 - Replace LocalGraphInternal::Info::lastSets vector by a unordered_map.

Benchmark process:
Firstly I gather xxx.temp.asm.js files from various sources. (Emscripten Tests, using EMCC_DEBUG=2)
Then process them to asm2wasm using -O3 parameter.
Run the application four times (for each case), keep the three best ones and compute the average.
The last column include remove seen set optimisation.

There are the results : 

| Test  | Base | Remove seen set | Replace vector -> umap |
| ------------- | ------------- | ------------- | ------------- |
| SQLite  | 12.969 s  | 7.789 s | 6.935 s | 
| Prolog | 81.139 s | 30.563 s | 28.505 s |
| Bullet | 1.458 s | 1.298 s | 1.286 s |
| Poppler | 11.913 s | 6.829 s | 5.973 s |
| Lua | 0.976 s | 0.893 s | 0.896 s |

Slighly different as I use Release version (instead of ReleaseWithDebugInfo) and times are provided directly by the application using std::chrono

I will test it next week on my company codebase to see the difference on another massive application.