@vloppin 2018-06-24
Add another one small optimisation.

 - Create a new struct for blocks containing only required information.
 - Use counter to avoid invalidation at each loop.

| Test  | Base | Previous | Latest |
| ------------- | ------------- | ------------- | ------------- |
| SQLite  | 12.969 s  | 6.935 s | 6.455 s |
| Prolog | 81.139 s | 28.505 s | 23.675 s |
| Bullet | 1.458 s | 1.286 s | 1.260 s |
| Poppler | 11.913 s | 5.973 s | 5.524 s |
| Lua | 0.976 s | 0.896 s | 0.891 s |