@vloppin 2018-06-25
> Btw, have you joined the w3c group yet?

In progress