@vloppin 2018-07-06
I'm still alive. I'm looking to join W3C group & legal stuff related to my employment contract. Sound good but it's requiring some times ... (Especially during summer holidays)