@vloppin 2018-07-20
My W3C registration is done ! So the last pending modification.